address comments: rename const file, return prev regexp
