+++ karma.conf.cjs
@@ -50,7 +50,7 @@ module.exports = function (config) {
             type: "asset/resource",
           },
           {
-            test: /\.(js|ts|tsx)$/,
+            test: /\.(js|tsx?)$/,
             loader: "ts-loader",
             exclude: /node_modules|\.d\.ts$/,
             options: { configFile: "tsconfig.karma.json" },
